refactor: pack buttons -> combine simple UI buttons into one import file
