--- src/components/Dashboard/PackCategory/TableButtons/AddCategoryButton.tsx
@@ -1,23 +1,0 @@
-import { Button, Icon } from 'semantic-ui-react';
-
-type AddCategoryButtonProps = {
-	onClick: () => void;
-};
-
-const AddCategoryButton = (props: AddCategoryButtonProps) => {
-	return (
-		<div>
-			<Button
-				color="grey"
-				basic
-				size="small"
-				className="add-item-table-button"
-				onClick={props.onClick}>
-				<Icon name="tree" />
-				Add Category
-			</Button>
-		</div>
-	);
-};
-
-export default AddCategoryButton;
